+++ packages/shared/src/data/config.ts
@@ -14,5 +14,5 @@ export default {
   defaultDuration: 343,
   minChunkSize: 5295308,
   loggerLevel: 1,
-  version: "2.0.8",
+  version: "2.0.9",
 } as ConfigSchema;